Organise all existing classes into packages
