--- src/main/java/Command.java
@@ -1,9 +1,0 @@
-public abstract class Command {
-
-    private String commandName;
-    public abstract void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException ;
-
-    public boolean isExit() {
-        return false;
-    }
-}
