+++ src/main/java/panav/command/EditCommand.java
@@ -1,4 +1,10 @@
-import java.io.IOException;
+package panav.command;
+
+import panav.exception.DukeException;
+import panav.exception.InvalidNumberException;
+import panav.storage.Storage;
+import panav.task.TaskList;
+import panav.ui.Ui;
 
 public class EditCommand extends Command {
 
@@ -26,6 +32,14 @@ public class EditCommand extends Command {
 
     }
 
+    /**
+     * Returns the index number for commands which manipulate the list.
+     *
+     * @param command The command which is manipulating list.
+     * @param len The number of elements in the list.
+     * @return Index number in command.
+     * @throws InvalidNumberException If the index doesn't exist.
+     */
     public static int readNumber(String command, int len) throws InvalidNumberException {
         int number = Integer.parseInt(String.valueOf(command.charAt(command.length() - 1)));
         if (number > len || number < 1) {
