+++ src/main/java/panav/command/EventCommand.java
@@ -1,3 +1,11 @@
+package panav.command;
+
+import panav.storage.Storage;
+import panav.task.Event;
+import panav.task.Task;
+import panav.task.TaskList;
+import panav.ui.Ui;
+
 public class EventCommand extends Command {
 
     public static final String COMMAND_WORD = "event";
