+++ src/main/java/panav/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package panav.command;
+
+import panav.storage.Storage;
+import panav.task.TaskList;
+import panav.ui.Ui;
+
 public class ExitCommand extends Command {
 
     public static final String COMMAND_WORD = "exit";
