+++ src/main/java/panav/exception/DukeException.java
@@ -1,3 +1,4 @@
+package panav.exception;
 public class DukeException extends Exception{
 
     public DukeException(String message) {
