+++ src/main/java/panav/exception/InvalidNumberException.java
@@ -1,3 +1,4 @@
+package panav.exception;
 public class InvalidNumberException extends DukeException{
 
     public InvalidNumberException(String message) {
