+++ src/main/java/panav/exception/ToDoDescriptionException.java
@@ -1,3 +1,4 @@
+package panav.exception;
 public class ToDoDescriptionException extends DukeException{
 
     public ToDoDescriptionException(String message) {
