+++ src/main/java/panav/task/Deadline.java
@@ -1,3 +1,4 @@
+package panav.task;
 public class Deadline extends Task{
     protected String by;
