+++ src/main/java/panav/task/Event.java
@@ -1,3 +1,4 @@
+package panav.task;
 public class Event extends Task{
     protected String from;
     protected String to;
