+++ src/main/java/panav/task/Task.java
@@ -1,6 +1,7 @@
+package panav.task;
 public class Task {
     protected String description;
-    protected boolean isDone;
+    public boolean isDone;
 
     public Task(String description) {
         this.description = description;
