+++ src/main/java/panav/task/ToDo.java
@@ -1,6 +1,10 @@
+package panav.task;
+
+import panav.exception.ToDoDescriptionException;
+
 public class ToDo extends Task{
 
-    public ToDo(String description) throws ToDoDescriptionException{
+    public ToDo(String description) throws ToDoDescriptionException {
         super(description);
         if (description.length() == 0) {
             throw new ToDoDescriptionException("☹ OOPS!!! The description of a todo cannot be empty.");
